Initial commit from fork
Adding workflow file which sets up a  docker container with a local version of the backend.
Then runs a few tests on this backend version which are also part of this commit. 
A line has been added in the docker-compose file to enable an alternate host name to use in the workflow